Move Account Add Function to add package
