+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/users/account/add/domain/account/Account.java
@@ -1,4 +1,4 @@
-package com.EmperorPenguin.SangmyungBank.api.users.account.domain.account;
+package com.EmperorPenguin.SangmyungBank.api.users.account.add.domain.account;
 
 import com.EmperorPenguin.SangmyungBank.api.users.add.domain.User.User;
 import lombok.Builder;
